Revert "Revert "(C)""
This reverts commit 860da28536e9ec80230bb182a0e4ed611a08a8bf.

+++ Client Applications/rti2rto/rti2rto.cpp
@@ -97,7 +97,7 @@ void ConvertRainbowTable(string sPathName, string sResultFileName, string sType)
 		return;
 	}
 	static CMemoryPool mp;
-	unsigned int nAllocatedSize;
+	uint64 nAllocatedSize;
 	BaseRTReader *reader = NULL;
 	if(sType == "RTI2")
 		reader = (BaseRTReader*)new RTI2Reader(sFileName);
@@ -108,7 +108,7 @@ void ConvertRainbowTable(string sPathName, string sResultFileName, string sType)
 		printf("Invalid table type '%s'", sType.c_str());
 		return ;
 	}
-	int size = reader->GetChainsLeft() * sizeof(RainbowChain);
+	uint64 size = reader->GetChainsLeft() * sizeof(RainbowChain);
 #ifdef _MEMORYDEBUG
 	printf("Starting allocation of %i bytes\n", size);
 #endif
